requirements.txt
akshare>=1.16.72
baostock>=0.8.8
backtrader>=1.9.76.123
openai>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
tabulate>=0.9.0
rich>=13.0.0
requests>=2.28.0
fastapi>=0.100.0
uvicorn>=0.22.0
websockets>=11.0
python-multipart>=0.0.6
easytrader>=0.23.0
httpx>=0.24.0
# Auth & User system
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
